Added more towns
